+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog for NeoFS Node
 ## [Unreleased]
 
 ### Added
+- `neofs-cli container policy check` command (#3790)
 
 ### Fixed
 - Unclosed compressed FSTree files (#3802)